# PERF: Optimize XmlRepair.cs